pyproject.toml
[project]
name = "xducs-mi24-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.28.0",
    "pillow>=11.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.4",
]

[tool.ruff]
line-length = 120
exclude = [
    "**/migrations/*",
]
